+++ packages/mind/src/generators/node-active.generator.ts
@@ -1,12 +1,4 @@
-import {
-    ACTIVE_STROKE_WIDTH,
-    PlaitBoard,
-    RectangleClient,
-    createG,
-    drawRoundRectangle,
-    getSelectedElements,
-    isSelectionMoving
-} from '@plait/core';
+import { ACTIVE_STROKE_WIDTH, PlaitBoard, RectangleClient, createG, drawRoundRectangle } from '@plait/core';
 import { MindElement, BaseData } from '../interfaces';
 import { getRectangleByNode } from '../utils/position/node';
 import { PRIMARY_COLOR } from '../constants/default';