+++ packages/mind/src/generators/node-collapse.generator.ts
@@ -1,6 +1,5 @@
 import { PlaitBoard, PlaitPointerType, Point, Transforms, createG, createText, drawLinearPath } from '@plait/core';
 import { MindElement, BaseData, PlaitMind, MindElementShape, LayoutDirection } from '../interfaces';
-import { AfterDraw, BaseDrawer } from '../base/base.drawer';
 import { getRectangleByNode } from '../utils/position/node';
 import { getShapeByElement } from '../utils/node-style/shape';
 import { EXTEND_OFFSET, EXTEND_DIAMETER } from '../constants/default';
@@ -12,8 +11,9 @@ import { filter, take } from 'rxjs/operators';
 import { getBranchColorByMindElement, getBranchWidthByMindElement } from '../utils/node-style/branch';
 import { getLayoutDirection, getPointByPlacement, moveXOfPoint, moveYOfPoint, transformPlacement } from '../utils/point-placement';
 import { HorizontalPlacement, PointPlacement, VerticalPlacement } from '../interfaces/types';
+import { AfterDraw, Generator } from '@plait/common';
 
-export class CollapseDrawer extends BaseDrawer implements AfterDraw {
+export class CollapseGenerator extends Generator<MindElement> implements AfterDraw {
     canDraw(element: MindElement<BaseData>): boolean {
         if (element.children.length && !PlaitMind.isMind(element)) {
             return true;
@@ -23,10 +23,7 @@ export class CollapseDrawer extends BaseDrawer implements AfterDraw {
 
     baseDraw(element: MindElement<BaseData>): SVGGElement {
         const collapseG = createG();
-        this.g = collapseG;
-
-        collapseG.classList.add('collapse-container');
-
+        collapseG.classList.add('collapse');
         const node = MindElement.getNode(element);
         const stroke = getBranchColorByMindElement(this.board, element);
         const branchWidth = getBranchWidthByMindElement(this.board, element);